+++ Raycasting/raycasting4.c
@@ -6,16 +6,16 @@
 /*   By: ojebbari <ojebbari@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/26 09:57:48 by ojebbari          #+#    #+#             */
-/*   Updated: 2024/03/27 13:53:29 by ojebbari         ###   ########.fr       */
+/*   Updated: 2024/03/27 15:09:34 by ojebbari         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../Cub3d.h"
 
-void	whileforvert(t_config *config, t_ray *ray, double nvtx, double nvty)
+void	whileforvert(t_config *config, t_ray *ray, float nvtx, float nvty)
 {
-	double	xtocheck;
-	double	ytocheck;
+	float	xtocheck;
+	float	ytocheck;
 
 	ray->wallhitxv = nvtx;
 	ray->wallhityv = nvty;
@@ -43,10 +43,10 @@ void	whileforvert(t_config *config, t_ray *ray, double nvtx, double nvty)
 
 void	cast_vertical_ray(t_config *config, t_ray *ray)
 {
-	double	nextverttouchx;
-	double	nextverttouchy;
-	double	yintercept;
-	double	xintercept;
+	float	nextverttouchx;
+	float	nextverttouchy;
+	float	yintercept;
+	float	xintercept;
 
 	xintercept = floor(config->player.x / (TILE_SIZE)) * (TILE_SIZE);
 	if (ray->israyfright)
@@ -66,10 +66,10 @@ void	cast_vertical_ray(t_config *config, t_ray *ray)
 	whileforvert(config, ray, nextverttouchx, nextverttouchy);
 }
 
-void	whileforhorz(t_config *config, t_ray *ray, double nhtx, double nhty)
+void	whileforhorz(t_config *config, t_ray *ray, float nhtx, float nhty)
 {
-	double	xtocheck;
-	double	ytocheck;
+	float	xtocheck;
+	float	ytocheck;
 
 	ray->wallhitxh = nhtx;
 	ray->wallhityh = nhty;
@@ -97,10 +97,10 @@ void	whileforhorz(t_config *config, t_ray *ray, double nhtx, double nhty)
 
 void	cast_horizontal_ray(t_config *config, t_ray *ray)
 {
-	double	nexthorztouchx;
-	double	nexthorztouchy;
-	double	yintercept;
-	double	xintercept;
+	float	nexthorztouchx;
+	float	nexthorztouchy;
+	float	yintercept;
+	float	xintercept;
 
 	yintercept = floor(config->player.y / (TILE_SIZE)) * (TILE_SIZE);
 	if (ray->israyfdown)
