+++ Raycasting/raycasting5.c
@@ -6,18 +6,18 @@
 /*   By: ojebbari <ojebbari@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 12:41:05 by ojebbari          #+#    #+#             */
-/*   Updated: 2024/03/27 13:58:56 by ojebbari         ###   ########.fr       */
+/*   Updated: 2024/03/27 15:09:34 by ojebbari         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../Cub3d.h"
 
-void	draw_up(t_config *config, double wallstripheight, int j, t_ray ray)
+void	draw_up(t_config *config, float wallstripheight, int j, t_ray ray)
 {
 	mlx_texture_t	*texture;
 	uint32_t		color;
 	int				i;
-	double			next_px;
+	float			next_px;
 	uint32_t		*pixels_tmp;
 
 	texture = set_texture(ray, config);
@@ -34,12 +34,12 @@ void	draw_up(t_config *config, double wallstripheight, int j, t_ray ray)
 	}
 }
 
-void	draw_down(t_config *config, double wallstripheight, int j, t_ray ray)
+void	draw_down(t_config *config, float wallstripheight, int j, t_ray ray)
 {
 	mlx_texture_t	*texture;
 	uint32_t		color;
 	int				i;
-	double			next_px;
+	float			next_px;
 	uint32_t		*pixels_tmp;
 
 	texture = set_texture(ray, config);
@@ -56,7 +56,7 @@ void	draw_down(t_config *config, double wallstripheight, int j, t_ray ray)
 	}
 }
 
-void	wall3d(t_config *config, double wallstripheight, int j, t_ray ray)
+void	wall3d(t_config *config, float wallstripheight, int j, t_ray ray)
 {
 	draw_up(config, wallstripheight, j, ray);
 	draw_down(config, wallstripheight, j, ray);
@@ -79,7 +79,7 @@ void	ceil2dfloor1d(t_config *config, int i)
 
 void	setup_wall(t_config *config)
 {
-	double	ray_distance;
+	float	ray_distance;
 	int		i;
 
 	i = 0;
